fix: move bcast new to behaviour
